text align float fix

+++ main.py
@@ -140,7 +140,7 @@ def create_quote_video(image_path, quotes, author):
         stroke_color='black',
         stroke_width=0.5,
         method='caption',
-        size=(bg_clip.w * 0.65, None),
+        size=(int(bg_clip.w * 0.65), None),
         text_align='center'
     ).with_duration(duration).with_position('center')
